Clean-up the KraftUtils class
