--- cluster-operator/src/test/java/io/strimzi/operator/cluster/model/KRaftUtilsTest.java
@@ -1,71 +1,0 @@
-/*
- * Copyright Strimzi authors.
- * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
- */
-package io.strimzi.operator.cluster.model;
-
-import io.strimzi.api.kafka.model.kafka.KafkaSpec;
-import io.strimzi.api.kafka.model.kafka.KafkaSpecBuilder;
-import io.strimzi.api.kafka.model.kafka.listener.GenericKafkaListenerBuilder;
-import io.strimzi.api.kafka.model.kafka.listener.KafkaListenerType;
-import io.strimzi.operator.common.model.InvalidResourceException;
-import io.strimzi.test.annotations.ParallelSuite;
-import io.strimzi.test.annotations.ParallelTest;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.is;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-@ParallelSuite
-public class KRaftUtilsTest {
-    @ParallelTest
-    public void testValidKafka() {
-        KafkaSpec spec = new KafkaSpecBuilder()
-                .withNewKafka()
-                    .withListeners(new GenericKafkaListenerBuilder()
-                            .withName("listener")
-                            .withPort(9092)
-                            .withTls(true)
-                            .withType(KafkaListenerType.INTERNAL)
-                            .build())
-                    .withNewKafkaAuthorizationOpa()
-                        .withUrl("http://opa:8080")
-                    .endKafkaAuthorizationOpa()
-                .endKafka()
-                .build();
-
-        assertDoesNotThrow(() -> KRaftUtils.validateKafkaCrForKRaft(spec));
-    }
-
-    @ParallelTest
-    public void testInvalidKafka() {
-        InvalidResourceException ex = assertThrows(InvalidResourceException.class, () -> KRaftUtils.validateKafkaCrForKRaft(null));
-        assertThat(ex.getMessage(), is("Kafka configuration is not valid: [The .spec section of the Kafka custom resource is missing]"));
-    }
-
-    @ParallelTest
-    public void testKRaftMetadataVersionValidation()    {
-        // Valid values
-        assertDoesNotThrow(() -> KRaftUtils.validateMetadataVersion("3.6"));
-        assertDoesNotThrow(() -> KRaftUtils.validateMetadataVersion("3.6-IV2"));
-
-        // Minimum supported versions
-        assertDoesNotThrow(() -> KRaftUtils.validateMetadataVersion("3.3"));
-        assertDoesNotThrow(() -> KRaftUtils.validateMetadataVersion("3.3-IV0"));
-
-        // Invalid Values
-        InvalidResourceException e = assertThrows(InvalidResourceException.class, () -> KRaftUtils.validateMetadataVersion("3.6-IV9"));
-        assertThat(e.getMessage(), containsString("Metadata version 3.6-IV9 is invalid"));
-
-        e = assertThrows(InvalidResourceException.class, () -> KRaftUtils.validateMetadataVersion("3"));
-        assertThat(e.getMessage(), containsString("Metadata version 3 is invalid"));
-
-        e = assertThrows(InvalidResourceException.class, () -> KRaftUtils.validateMetadataVersion("3.2"));
-        assertThat(e.getMessage(), containsString("The oldest supported metadata version is 3.3-IV0"));
-
-        e = assertThrows(InvalidResourceException.class, () -> KRaftUtils.validateMetadataVersion("3.2-IV0"));
-        assertThat(e.getMessage(), containsString("The oldest supported metadata version is 3.3-IV0"));
-    }
-}
